+++ CHANGELOG.md
@@ -14,7 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Allow modules to alter dashboard panes #868](https://github.com/farmOS/farmOS/pull/868)
 - [Add geometry/location fields to CSV importers #815](https://github.com/farmOS/farmOS/pull/815)
 - [Add an asset.logs service for retrieving logs that reference an asset #850](https://github.com/farmOS/farmOS/pull/850)
-- [Add farmOS-timeline library #862](https://github.com/farmOS/farmOS/pull/862)
+- [Add farmOS-timeline library (experimental) #862](https://github.com/farmOS/farmOS/pull/862)
 
 ### Changed