+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Add geometry/location fields to CSV importers #815](https://github.com/farmOS/farmOS/pull/815)
 - [Add an asset.logs service for retrieving logs that reference an asset #850](https://github.com/farmOS/farmOS/pull/850)
 - [Add farmOS-timeline library (experimental) #862](https://github.com/farmOS/farmOS/pull/862)
+- [Build multi-platform Docker images/manifests #864](https://github.com/farmOS/farmOS/pull/864)
 
 ### Changed