+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.3.0] 2024-09-24
+
 ### Added
 
 - [Add bulk delete action to farm_log_quantity View #860](https://github.com/farmOS/farmOS/pull/860)
@@ -820,7 +822,8 @@ moving forward.
     Drupal 7, which required a complete refactor of the codebase. By comparison,
     updating from Drupal 9 to 10 will simply involve updating deprecated code.
 
-[Unreleased]: https://github.com/farmOS/farmOS/compare/3.2.4..HEAD
+[Unreleased]: https://github.com/farmOS/farmOS/compare/3.3.0..HEAD
+[3.2.4]: https://github.com/farmOS/farmOS/releases/tag/3.3.0
 [3.2.4]: https://github.com/farmOS/farmOS/releases/tag/3.2.4
 [3.2.3]: https://github.com/farmOS/farmOS/releases/tag/3.2.3
 [3.2.2]: https://github.com/farmOS/farmOS/releases/tag/3.2.2